Add certificate principals display for hiba-gen
Now that we can use __SELF__ to restrict the target roles, it can be
useful to have a debug mechanism to show thelist of principals withing a
certificate.

Also, automatically set plural/singular when showing the count of principal(s)
and grant(s) based on the number contained in the certificate.