Add certificate principals display for hiba-gen
Now that we can use @PRINCIPALS to restrict the target roles, it can be
useful to have a debug mechanism to show the list of principals within a
certificate.

Also, automatically set plural/singular when showing the count of principal(s)
and grant(s) based on the number contained in the certificate.