Add explicit function for serializing format.
HIBA extensions can be stored raw using SSH Buffers (rfc4251#section-5)
and PROTOCOL.extensions, or base64 encoded. The default up until now was
to serialize an extension and automatically apply base64 encoding, but
this is not always desirable.

Instead, two functions are now provided with a macro providing backward
compatibility.