Add support for compressed HIBA grant extensions.
Compression is optional and support conditional on a configure option:
--enable-extension-compression in order to make the dependency on zlib
optional. Requesting compressed extensions when support for compression
is not built-in will result in transparently ignoring the request and
return uncompressed but valid data.

Compression can be applied on top of all extension formats (single
grant, multi grants, base64, or raw), but provides best result when
applied on top of raw format.

The hibaext_encode() macro is updated to provie backward compatibility
for previous users and default to using uncompressed format.