Add support for negative matching constraints.
This negative matching constraints are expressed as: "not a key which
value is X".

This is implemented using the `!` as a key modifier. The documentation
is updated to reflect this new capability in the HIBA protocol.

A new set of regression tests is also provided to confirm the behavior
of negative matching constraints.

The current version of the protocol is bumped to 2 since negative
matching potentially introduces backward incompatible grants.

At generation time, a grant will have a min required HIBA parser version
of 1 (since the positive match is backward compatible), and the min
required HIBA parser version will be bumped to 2 if a negative matching
constraint key is added, which will cause any older HIBA binary to
simply refuse the grant.