Fix ubsan warning on null pointer usage
For the allocation of a new revocation entry, hiba-grl passes a NULL
pointer as source of memcpy(), triggering a ubsan warning.
Since it also passes a size of zero, this did not actually trigger any
error condition, but should be fixed nonetheless.

The warning in question is:
revocations.c:439:3: runtime error: null pointer passed as argument 2, which is declared to never be null