Improve support for quotes in 'options' constraint.
The options constraint in a HIBA grant can be used, for example, for
restricting users to a given command similar to authorized_keys or
authorized_users.

According to OpenSSH specifications, in this case, the command can
contain quotes by escaping them with a backslash, which was not
implemented in the sanity check until now.