include sys/types.h before log.h
OpenSSH 10.0 has some declarations in log.h that use the type `u_int`,
which is usually provided by the platform's sys/types.h.

Including OpenSSH's `defines.h` might be a better fix, because it
includes replacement types when the system's headers lack stuff, but
that's a bit more risky to include as it replaces a bunch of stuff.